Add finance & reporting module: expenses, advances, reports
- Add domain entities for expenses, cash advances, and monthly financial reports with business logic
- Extend enums for inventory movement and cash advance statuses
- Track unit cost in inventory movement items for material expense accounting
- Update DbContext and ERD for new finance and shipment tables
- Implement CRUD and validation for expenses and cash advances, enforcing user permissions
- Add commands/queries for calculating, approving, closing, and listing monthly financial reports
- Integrate expense and advance data into financial reporting
- Use FluentValidation for input validation throughout
- Minor DTO and command consistency fixes; comment out obsolete DbContext code

+++ src/MyFactory.Application/Features/Products/UpdateProduct/UpdateProductCommand.cs
@@ -6,6 +6,6 @@ namespace MyFactory.Application.Features.Products.UpdateProduct;
 public sealed record UpdateProductCommand(
     Guid ProductId,
     string Name,
-    decimal? PlanPerHour,
+    int? PlanPerHour,
     ProductStatus Status
 ) : IRequest<Guid>;
